Rename dispatcher file to invokeHooks
